+++ lib/base/types.dart
@@ -1,25 +1,24 @@
 // TODO separate types into different files
 
 /// An [Store] records all events happened in app and acts as the single source of truth.
-abstract class Store<E extends Object>
-    implements Projectable<E>, Publishable<E> {
+abstract class Store implements Projectable, Publishable {
   Unsubscribe subscribe(Subscriber subscriber);
 }
 
 /// [Store] interface for store enhancers.
-abstract class StoreForEnhancer<E extends Object> extends Store<E> {
+abstract class StoreForEnhancer extends Store {
   int get cursor;
-  List<E> get events;
-  void replaceEvents(List<E> events, [int cursor]);
+  List get events;
+  void replaceEvents(List events, [int cursor]);
 }
 
-abstract class Projectable<E extends Object> {
-  P project<P>(Projector<P, E> projector);
+abstract class Projectable {
+  P project<P>(Projector<P> projector);
 }
 
 /// [Store.publish] interface
-abstract class Publishable<E extends Object> {
-  E publish(E event);
+abstract class Publishable {
+  E publish<E>(E event);
 }
 
 /// Callback when [Store] is updated.
@@ -29,16 +28,14 @@ typedef Subscriber = void Function();
 typedef Unsubscribe = void Function();
 
 /// Store creator for middleware.
-typedef StoreCreator<E extends Object> = StoreForEnhancer<E> Function(
-    List<E> prepublish);
+typedef StoreCreator = StoreForEnhancer Function(List prepublish);
 
 /// [StoreEnhancer] are called "store enhancer" in Redux.
 /// They wrap store and enable features like time travel.
-typedef StoreEnhancer<E extends Object> = StoreCreator<E> Function(
-    StoreCreator<E> inner);
+typedef StoreEnhancer = StoreCreator Function(StoreCreator inner);
 
 /// A [Projector] calculates view derived from events.
 /// [prev] is nullable
 /// Hint: an object with a `call()` method will also work.
-typedef Projector<P extends Object, E extends Object> = P Function(
-    P prev, List<E> event, Projectable<E> projectable);
+typedef Projector<P extends Object> = P Function(
+    P prev, List event, Projectable projectable);