+++ test/base/create_store_test.dart
@@ -7,12 +7,12 @@ import '../test_utils.dart';
 void main() {
   group('createStore', () {
     test('should return a valid EventStore', () {
-      final Store<MathEvent> s = createStore<MathEvent>([]);
-      expect(s, Matcher.TypeMatcher<Store<MathEvent>>());
+      final Store s = createStore(<Object>[]);
+      expect(s, Matcher.TypeMatcher<Store>());
     });
 
     test('should dispatch event to subscriber', () {
-      final Store<MathEvent> s = createStore<MathEvent>([]);
+      final Store s = createStore(<Object>[]);
       var value = 0;
       final unsubscribe = s.subscribe(() {
         value += 1;
@@ -28,7 +28,7 @@ void main() {
     });
 
     test('should return projection', () {
-      final Store<MathEvent> s = createStore<MathEvent>([]);
+      final Store s = createStore(<Object>[]);
       s.publish(Minus('1'));
       s.publish(Minus('2'));
       s.publish(Plus(3));
@@ -38,12 +38,10 @@ void main() {
     });
 
     test('should cache state result for the same projector', () {
-      final Store<MathEvent> s = createStore<MathEvent>([]);
+      final Store s = createStore(<Object>[]);
       var projectCount = 0;
-      final Projector<int, MathEvent> projector =
-          (prev, events, _) => projectCount++;
-      final Projector<int, MathEvent> projector2 =
-          (prev, events, _) => projectCount++;
+      final Projector<int> projector = (prev, events, _) => projectCount++;
+      final Projector<int> projector2 = (prev, events, _) => projectCount++;
       s.publish(Minus('1'));
       s.publish(Minus('2'));
       s.publish(Plus(3));
@@ -61,10 +59,9 @@ void main() {
     });
 
     test('should clean state cache after events got replaced', () {
-      final Store<MathEvent> s = createStore<MathEvent>([]);
+      final Store s = createStore(<Object>[]);
       var projectCount = 0;
-      final Projector<int, MathEvent> projector =
-          (prev, events, _) => projectCount++;
+      final Projector<int> projector = (prev, events, _) => projectCount++;
       s.publish(Minus('1'));
       s.publish(Minus('2'));
       s.publish(Plus(3));
@@ -74,7 +71,7 @@ void main() {
       expect(projectCount, 2);
       s.project(projector);
       expect(projectCount, 2);
-      (s as StoreForEnhancer<MathEvent>).replaceEvents([], 0);
+      (s as StoreForEnhancer).replaceEvents(<Object>[], 0);
       s.project(projector);
       expect(projectCount, 3);
       s.project(projector);