+++ test/enhancers/with_side_effect_test.dart
@@ -8,7 +8,7 @@ import '../test_utils.dart';
 void main() {
   group('withSideEffect', () {
     test('should be able to transform events', () async {
-      final store = createStore<MathEvent>([], [withSideEffect(add1When3)]);
+      final store = createStore(<Object>[], [withSideEffect(add1When3)]);
       store.publish(Plus(1));
       store.publish(Plus(1));
       await Future(() {});
@@ -20,8 +20,7 @@ void main() {
   });
 }
 
-Stream<MathEvent> add1When3(
-    Stream<MathEvent> events, Projectable<MathEvent> store) async* {
+Stream<MathEvent> add1When3(Stream events, Projectable store) async* {
   await for (final event in events) {
     if (store.project(sum) == 3) {
       yield Plus(1);